fix(bus, dram, riscv): change address type from unsigned int to unsigned long for consistency

+++ src/bus.c
@@ -1,9 +1,9 @@
 #include "bus.h"
 
-u32* __fastcall__ busLoad(struct Bus* bus, unsigned int address) {
+u32* __fastcall__ busLoad(struct Bus* bus, unsigned long address) {
     return drmLoad(&bus->dram, address);
 }
 
-void __fastcall__ busStore(struct Bus* bus, unsigned int address, u32* value) {
+void __fastcall__ busStore(struct Bus* bus, unsigned long address, u32* value) {
     drmStore(&bus->dram, address, value);
 }
