+++ src/bus.h
@@ -8,8 +8,8 @@ struct Bus {
     struct DRAM dram;
 };
 
-u32* __fastcall__ busLoad(struct Bus* bus, unsigned int address);
+u32* __fastcall__ busLoad(struct Bus* bus, unsigned long address);
 
-void __fastcall__ busStore(struct Bus* bus, unsigned int address, u32* value);
+void __fastcall__ busStore(struct Bus* bus, unsigned long address, u32* value);
 
 #endif
